pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
    "setuptools",
    "wheel",
]

[project]
name = "patchelfy"
description = "Add code in your ELF without breaking anything"
version = "0.1"
authors = [{name = "Edoardo"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pwntools",
    "capstone",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: Unix",
]

[project.urls]
"Homepage" = "https://github.com/Angelo942/patchelfy"
"Bug Tracker" = "https://github.com/Angelo942/patchelfy/issues"